+++ src/Traits/HelperExceptionTrait.php
@@ -39,4 +39,35 @@ trait HelperExceptionTrait
 
         return json_encode($exceptionArray, JSON_THROW_ON_ERROR | static::jsonFlags());
     }
+
+
+    /**
+     * Возвращает исключение в виде массива
+     * @see ../../tests/HelperExceptionTrait/HelperExceptionToArrayTest.php
+     *
+     * @param ?Throwable $value
+     * @return array
+     */
+    public static function exceptionToArray(?Throwable $value): array
+    {
+        if (is_null($value)) {
+            return [];
+        }
+
+        return [
+            'code' => $value->getCode(),
+            'message' => $value->getMessage(),
+            'exception' => static::pathClassName($value::class),
+            'file' => static::stringConcat(
+                ':',
+                trim(static::stringReplace($value->getFile(), static::pathRoot(), '', useRegexp: false), '/'),
+                $value->getLine(),
+            ),
+            'trace' => static::arrayExcludeTraceVendor($value->getTrace()),
+            ...($value::class === 'Illuminate\Database\QueryException'
+                ? ['sql' => static::stringDeleteMultiples(static::stringReplace($value->getRawSql(), ['"' => '']), ' ')]
+                : []
+            ),
+        ];
+    }
 }